backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
requests==2.31.0
reportlab==4.0.7
PyPDF2==3.0.1
pdfplumber==0.10.3
openpyxl==3.1.5
python-docx==1.1.0
# Embeddings
sentence-transformers==2.2.2
# Optional: for OpenAI embeddings (uncomment if using OpenAI provider)
# openai==1.12.0
# Optional: for Vertex AI embeddings (uncomment if using Vertex provider)
# google-cloud-aiplatform==1.38.0

# Vector store
pinecone>=4.0.0